Add Option type
Went with a class-based implementation to make it comparable to
Promises. I tried to get the typing right, but didn't manage in
a reasonable amount of time.

Some.value is of type T and None.value is of type never. Just what
I want. Sadly, Option.value is not that strict. Next time.

Now even more so than ever, I prefer the collection of pure functions
used in FP-TS and F#. So much more straight foreward. But that
way of constructing programmes only works so well in JS.

Used the F# docs to verify the bind signature:
https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#bind

Got some class typing help from StackOverflow: (what a pain that was)
https://stackoverflow.com/questions/70596483/implementing-maybe-option-type-in-typescript-using-class-inheritance