Add function-based Option type
Doing so for fun and profit. Well, more for contrast.

The implementation feels nicer. The union of Some and None
is explicit and getting the value requires no hoops.

But without pipes |>  their usage sucks, as several consequitive
maps or binds don't work well. See the tests for examples.

I guess the class version works better in JS :(