Add option evaluation
Use Option types like you would promises in async/await. Similar
to computational expressions in F#.

Within an evaluation, yielding an Option returns the unwrapped value.
If the Option is None, the whole thing exits and returns a None. No other
code is executed.

This does not yet fully copy the async/await/promise thing, as
Promise.then() are both a map and a bind.