Make value accessible only in Some
Previously, both Some and None could access the value and
in the case of the latter an error was thrown. The types weren't
very nice either, as .value was awkwardly typed as T | never, no
matter in what state the Option was in.

This moves the  value getter to the Some class and update the tests.
The @ts-expect-error comment will cause a TS error if value exists.

None can no longer access the non-existent value, cleaning up the types
and removing the need to throw an error.